runtime: inline stack pool owner module
