+++ .github/workflows/microcrate-ci.yml
@@ -67,7 +67,7 @@ jobs:
               concurrency) match '^crates/concurrency' && v=true ;;
               governance) match '^(crates/governance|crates/bdd|tests/governance|crates/feature-policy|crates/runtime-governance)' && v=true ;;
               bdd) match '^crates/bdd' && v=true ;;
-              parser) match '^(crates/parser-|crates/common-syntax-core|crates/linecol-core|crates/stack-pool-core|crates/parsetable-metadata|glr-core/|tablegen/)' && v=true ;;
+              parser) match '^(crates/parser-|crates/common-syntax-core|crates/linecol-core|crates/parsetable-metadata|glr-core/|tablegen/)' && v=true ;;
               core) match '^(ir/|glr-core/|tablegen/|common/)' && v=true ;;
               runtime) match '^(runtime/|runtime2/|macro/|tool/|cli/)' && v=true ;;
             esac
@@ -180,7 +180,6 @@ jobs:
             adze-parser-contract
             adze-parser-governance-contract
             adze-linecol-core
-            adze-stack-pool-core
             adze-parsetable-metadata
             adze-common-syntax-core
             adze-concurrency-caps-contract-core
@@ -320,7 +319,6 @@ jobs:
             -p adze-parser-governance-contract \
             -p adze-common-syntax-core \
             -p adze-linecol-core \
-            -p adze-stack-pool-core \
             -p adze-parsetable-metadata \
             -- --test-threads=$RUST_TEST_THREADS
