+++ Cargo.toml
@@ -14,7 +14,6 @@ members = [
   "crates/runtime-governance-api",
   "crates/runtime-governance-matrix",
   "crates/runtime2-governance",
-  "crates/stack-pool-core",
   "crates/governance-matrix-contract",
   "crates/governance-matrix-core-impl",
   "crates/governance-matrix-core",
