+++ docs/reference/PUBLISH_CHECKLIST.md
@@ -25,11 +25,10 @@ the core crates that depend on them. The full set (in order) is:
 1. All `adze-concurrency-*` crates (caps-contract-core, map-core, env-core,
    init-core, caps-core, etc.)
 2. `adze-linecol-core`
-3. `adze-stack-pool-core`
-4. `adze-common-syntax-core`
-5. `adze-bdd-scenario-core` -> `adze-bdd-grid-core`
-6. `adze-governance-metadata` -> `adze-parsetable-metadata`
-7. `adze-runtime-governance` -> `adze-runtime-governance-api`
+3. `adze-common-syntax-core`
+4. `adze-bdd-scenario-core` -> `adze-bdd-grid-core`
+5. `adze-governance-metadata` -> `adze-parsetable-metadata`
+6. `adze-runtime-governance` -> `adze-runtime-governance-api`
 
 ## Pre-publish verification
