+++ runtime/Cargo.toml
@@ -71,7 +71,6 @@ adze-linecol-core = { version = "0.1.0", path = "../crates/linecol-core" }
 ts_bridge = { package = "ts-bridge", version = "0.1.0", path = "../tools/ts-bridge", optional = true, default-features = false }
 tree_sitter_json = { package = "tree-sitter-json", version = "0.24", optional = true }
 adze-runtime-governance-api = { version = "0.1.0", path = "../crates/runtime-governance-api" }
-adze-stack-pool-core = { version = "0.1.0", path = "../crates/stack-pool-core" }
 
 [dev-dependencies]
 insta = "1.43.1"
